Fixed bug when scaling ProgressIndicator, it is now proportional scale. Changed Rectangle to Ellipse to match Win8 Styling.
Set MinHeight and MaxHeight for ProgressIndicator. Changed the MetroCircleButtonStyle to change the opacity of the content property when the IsEnabled=False.

Changed TabItem Headers to use HeaderFamily instead of DefaultFont

Fixed Window Controls from not turning White when the dark theme is used and the titlebar banner is off.

Added NormalForegroundBrush which uses NormalForegroundColor
Changed BaseDark NormalForegroundColor from White to GrayBrush6
Updated ListView, Expander, GroupItem,  to have more of a Metro feel, uses NormalForegroundColor for content.
Added 2 font families used in Metro apps. Segoe Semilight and Semibold and changed the columnheader and expander to those fonts and recommended size.