Fixed popup event handling bug
Previously the owning window's PreviewMouseDown event, used to auto close the popover, was being subscribed to
on the popover's load event and unsubscribed on unload however this didn't take into account that the Load could be called multiple
times. The popover now keeps track of if it's subscribe to the given event using a disposable handler, which is also used
to unsubscribe.