Fixed autoclose no longer closes popover on selecting Combo Box item.
Previously clicking a dropdown item would cause the containing popover
to close if it's AutoClose is set to True (default). This occured because the
combobox uses a popup to display the dropdown which is outside the
popovers visual tree which meant the popover thought that the user had
clicked elsewhere in the window. This is fixed by using DependancyObject.GetParentObject()
which falls back to the logical tree when required.