Fix IsReadonly issue
[Fix] If in NumericUpDown the property IsReadOnly is set to true, then changing the values by clicking + and - Buttons is still possible
As you can see in https://github.com/MahApps/MahApps.Metro/blob/master/MahApps.Metro/Themes/NumericUpDown.xaml#L120 the only that disables the change of the value is the Trigger. Now even if the Style is customized and this kind of Trigger is not set, then the Change of value is not possible.