If the user.config file is corrupt (e.g. empty) the whole application crashes
I had the problem, that a customer could suddenly not open my
application anymore. It turned out that this is because the user.config
file was just empty and this caused a ConfigurationException. Then I
added an UnhandledExceptionHandler but this exception seems to happen
earlier. The only working solution ist to delete the odd user.config and
restart the application.
For more details see here:

http://www.codeproject.com/Articles/30216/Handling-Corrupt-user-config-Settings