[RFC] Create Accents based on user input
The ThemeManager has a got a new function that takes a userdefined color and creates a new Accent from it. The Accent will be created for all base themes, so you can switch to light and dark as usual.

Please let me know, what you think about this :-)

You can test it in the demo app under accents-menu.