Add new DP UpDownButtonsFocusable
Add new dependency property UpDownButtonsFocusable to allow changing the Focusable property of the up and down buttons.

The default is true to keep the  current behavior. If the up or down button will be used then the focus will move to it and triggers the lost focus event of the inner TextBox to change the value.