(GH-3533) Collapsed SplitView Pane not shown when set Visibility
When a SplitView was initialized with Visibility=Visibility.Collapsed then the Pane does not appear when changing to Visibility=Visibility.Visible.

If the SplitView starts invisible then the inner template part PaneClipRectangle is not available at the loaded event and changing the visual state has no effect.

To fix this the loaded event with the visual state change must be called when the template will be applied. If the SplitView is already loaded then this can be called directly.