Update DropDownButton.cs
Allow using the DropDownButton without a dropdown menu.

If it is necessary to use the DropDownButton without a menu under some circumstances and the "Click" event is used, the button will still try to open the empty menu, which can be seen by the user as a thick black line.
This change only tries to open the menu if there actually are some items inside.