(#4157) Fix BindingExpression errors to SelectedHSVColor and SelectedColor
These errors occurs because of the Color binding on the SolidColorBrush.

Binding with default data source as DataContext wont work for SolidColorBrush type as they are not framework elements.

It is a DependencyObject, and Color is a DependencyProperty, so Bindings on it work fine - until you're inside a context where DataContext isn't passed.