fix(dialogs): #2400 fix using access keys (mnemonics) in dialogs
There was a problem with access keys (mnemonics). Pressing an access key that is assigned to a control in the MetroWindow will switch keyboard focus to the MetroWindow, while a dialog was shown. Furthermore the access keys of the dialog buttons are not working if the same access keys are used for a control inside the MetroWindow.