fix: #4124 fix scrolling and width of RichTextBox
Remove the workaround for Documents PageWidth. The main problem here is that a RichTextBox inside a ScrollViewer with horizontal scrolling is still broken. The RichTextBox will wrap the text after each letter.

To solve this the ScrollViewer binds now the Width to the parent ActualWidth of a Grid.

Additional the wrong Margin of the ScrollViewer host for all text controls will be fixed by using the new attached property `ScrollViewerHelper.ScrollContentPresenterMargin`.