fix: #4257 don't change NUD value on lost focus event
The NUD has a strange behavior if it's located inside a TabItem and the value is bound to a MVVM property which is used then in all TabItems (via DataTemplate).

So updating the value from the textbox is not a good way. Instead it should only update the text inside the textbox because the value is now converted on the fly when the text is changed by the user or the up down buttons.

BREAKING CHANGE: remove ChangeValueOnTextChanged property

The ChangeValueOnTextChanged is removed, because of the ugly lost focus problem. So after all it should now work and convert the text on the fly to it's value.